Changes:  - merged changes between creation of the branch and HEAD for:     - build system changes     - 3rd-party libraries     - atermpp and lts libraries     - some tools that do not directly depend on the data library
   Until after the remaining changes to tools and data, core, lps and pbes
   libraries and tools are merged the branch will only compile with squadt-support
   disabled.



git-svn-id: https://svn.win.tue.nl/repos/MCRL2/branches/new-data-lib@5568 21252fe5-60e8-0310-8ff0-c1ae1001594e

+++ tools/diagraphica/attrdiscr.cpp
@@ -8,6 +8,8 @@
 //
 /// \file ./attrdiscr.cpp
 
+#include "wx.hpp" // precompiled headers
+
 #include "attrdiscr.h"
 
 // -- constructors and destructor -----------------------------------
@@ -156,8 +158,8 @@ void AttrDiscr::clusterValues(
 
 // -----------------------
 void AttrDiscr::moveValue(
-    const int &idxFr,
-    const int &idxTo )
+    const size_t &idxFr,
+    const size_t &idxTo )
 // -----------------------
 {
     try
@@ -168,7 +170,7 @@ void AttrDiscr::moveValue(
         if ( idxFr < idxTo )
         {
             // move all values after idxFr 1 pos up
-            for ( int i = idxFr; i < idxTo; ++i )
+            for ( size_t i = idxFr; i < idxTo; ++i )
             {
                 curValues[i] = curValues[i+1];
                 curValues[i]->setIndex( i );
@@ -180,7 +182,7 @@ void AttrDiscr::moveValue(
         else if ( idxTo < idxFr )
         {
             // move all values before idxFr 1 pos down
-            for ( int i = idxFr; i > idxTo; --i )
+            for ( size_t i = idxFr; i > idxTo; --i )
             {
                 curValues[i] = curValues[i-1];
                 curValues[i]->setIndex( i );
